@@ -1,38 +1,37 @@
 {
-  "name": "istekram",
+  "name": "client2",
   "version": "0.1.0",
   "private": true,
-  "homepage": "https://ybaha.github.io/istekram/",
   "dependencies": {
+    "@testing-library/jest-dom": "^5.16.2",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
     "@types/firebase": "^3.2.1",
-    "@types/node": "^16.6.1",
-    "@types/react": "^17.0.17",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-timeago": "^4.1.3",
-    "axios": "^0.21.1",
-    "firebase": "^8.9.1",
-    "mobx": "^6.3.3",
-    "mobx-react-lite": "^3.2.1",
-    "node-sass": "^4.14.1",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^16.11.22",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
+    "axios": "^0.25.0",
+    "firebase": "^9.6.6",
+    "mobx": "^6.3.13",
+    "mobx-react-lite": "^3.2.3",
+    "node-sass": "^7.0.1",
     "react": "^17.0.2",
     "react-crossfade-image": "^1.2.0",
     "react-dom": "^17.0.2",
-    "react-icons": "^4.2.0",
-    "react-images-uploading": "^3.1.3",
-    "react-indiana-drag-scroll": "^1.8.1",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.0",
+    "react-icons": "^4.3.1",
+    "react-indiana-drag-scroll": "^2.1.0",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "5.0.0",
     "react-timeago": "^6.2.1",
-    "styled-components": "^5.3.1",
-    "typescript": "~4.0.5",
-    "web-vitals": "^0.2.4"
+    "styled-components": "^5.3.3",
+    "typescript": "^4.5.5",
+    "web-vitals": "^2.1.4"
   },
   "scripts": {
-    "start": "cross-env REACT_APP_SERVER=http://localhost:5000/ react-scripts start",
+    "start": "cross-env REACT_APP_SERVER=http://localhost:3001/ react-scripts start",
     "build": "cross-env REACT_APP_SERVER=https://istekram.herokuapp.com/ react-scripts build",
     "dev": "cross-env REACT_APP_SERVER=https://istekram.herokuapp.com/ react-scripts start",
-    "test": "craco test",
-    "eject": "craco eject",
     "predeploy": "npm run build",
     "deploy": "gh-pages -d build"
   },
@@ -44,26 +43,24 @@
   },
   "browserslist": {
     "production": [
-      ">0.3%",
-      "not ie 11",
+      ">0.2%",
       "not dead",
       "not op_mini all"
     ],
     "development": [
-      ">0.3%",
-      "not ie 11",
-      "not dead",
-      "not op_mini all"
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
     ]
   },
   "devDependencies": {
-    "@types/react-router-dom": "^5.1.8",
-    "@types/styled-components": "^5.1.14",
-    "autoprefixer": "^9.8.6",
+    "@types/react-timeago": "^4.1.3",
+    "@types/styled-components": "^5.1.21",
+    "autoprefixer": "^10.4.2",
     "cross-env": "^7.0.3",
     "gh-pages": "^3.2.3",
-    "postcss": "^7.0.36",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.14"
+    "postcss": "^8.4.6",
+    "tailwindcss": "^3.0.18"
   },
   "resolutions": {
     "styled-components": "^5"
