@@ -1,24 +1,12 @@
 import React from "react";
-import { Route, Redirect } from "react-router-dom";
+import { Route, Navigate, Outlet } from "react-router-dom";
 import { useAuth } from "./AuthContext";
 
-const PrivateRoute: React.FC<{ [x: string]: any }> = ({
-  Component,
-  ...rest
-}) => {
+const PrivateRoute: React.FC = ({ children }): any => {
   const { currentUser } = useAuth();
-  return (
-    <Route
-      {...rest}
-      render={(props) => {
-        return currentUser ? (
-          <Component {...props} />
-        ) : (
-          <Redirect to="/istekram/login" />
-        );
-      }}
-    ></Route>
-  );
+  if (!currentUser) return <Navigate to="/istekram/login" />;
+
+  return children;
 };
 
 export default PrivateRoute;
