@@ -1,11 +1,11 @@
 import React from "react";
 import { DDProfile, DDSaved, DDSettings, DDSwitch } from "../icons";
 import { useAuth } from "../../Firebase/AuthContext";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 export default function Dropdown() {
   const { logout, getCurrentUsername } = useAuth();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   return (
     <div
@@ -24,7 +24,7 @@ export default function Dropdown() {
       <div
         className="dd-item"
         onClick={() => {
-          history.push("/istekram/" + getCurrentUsername());
+          navigate("/istekram/" + getCurrentUsername());
         }}
       >
         <DDProfile />
