@@ -1,9 +1,9 @@
-import firebase from "firebase/app";
-import "firebase/auth";
-import "firebase/database";
-import "firebase/storage";
+import { initializeApp } from "firebase/app";
+import { getAuth } from "firebase/auth";
+import { getDatabase } from "firebase/database";
+import { getStorage } from "firebase/storage";
 
-const app = firebase.initializeApp({
+const app = initializeApp({
   apiKey: "AIzaSyDANWNmCkloLJqH7xCwSt6UFKvAFE2eDvs",
   authDomain: "istekram-5e37a.firebaseapp.com",
   databaseURL:
@@ -15,7 +15,7 @@ const app = firebase.initializeApp({
   measurementId: "G-GDZ395F31B",
 });
 
-export const auth = app.auth();
-export const db = app.database();
-export const st = app.storage();
+export const auth = getAuth(app);
+export const db = getDatabase(app);
+export const st = getStorage(app);
 export default app;
